ADCM-3078 refactor audit
